+++ src/util.rs
@@ -5,7 +5,11 @@ use std::slice;
 use binding::global::RubySpecialConsts;
 use types::{c_char, c_int, c_void, Argc, InternalValue, Value};
 
-use {AnyObject, Object};
+use {AnyObject, Object, Boolean};
+use crate::rubysys::rproc::{
+    rb_obj_is_proc,
+    rb_obj_is_method,
+};
 
 pub unsafe fn cstr_to_string(str: *const c_char) -> String {
     CStr::from_ptr(str).to_string_lossy().into_owned()
@@ -98,4 +102,12 @@ where
 
 pub unsafe fn ptr_to_data<R>(ptr: *mut c_void) -> R {
     *Box::from_raw(ptr as *mut R)
+}
+
+pub fn is_proc(obj: AnyObject) -> bool {
+    Boolean::from(unsafe { rb_obj_is_proc(obj.value()) }).to_bool()
+}
+
+pub fn is_method(obj: AnyObject) -> bool {
+    Boolean::from(unsafe { rb_obj_is_method(obj.value()) }).to_bool()
 }